Either record or show heatmap data

By default, process_heatmap records the heatmap. However, if you browse to the app url and add ?get_heatmap query string, the function will run in display mode to see the heatmap data.

Usage

process_heatmap(...)

Arguments

...

Pass in parameters for record_heatmap and download_heatmap.

Note

Specify options("shinyHeatmap.debug" = TRUE) to get useful debugging messages, particularly when setting up your heatmap.