Record heatmap data to a file

On the JS side, Shiny creates a input$heatmap_data containing the current click data to be recorded to the file

Usage

record_heatmap(
  trigger = NULL,
  path = "www",
  target = ".container-fluid",
  type = c("click", "move"),
  timeout = 10,
  session = shiny::getDefaultReactiveDomain(),
  ...
)

Arguments

trigger

Reactive trigger to initialized the heatmap recording. This is useful if your app contains tabs or navbar.

path

Previously saved heatmap data for persistence.

target

Container selector hosting the heatmap canvas. Default to Shiny fluidPage container. Be careful to change it if you use another template.

type

Event type: click or mouse move. Default to click.

timeout

Necessary if the page needs time to load. Expressed in milliseconds.

session

Shiny session object. Useful to store heatmap data.

...

Internal. Don't use.